Move entitlements selector to the moves file.
